Remove natsio and nitxo clients
